restore real middlebox creation

+++ src/main/java/net/floodlightcontroller/nfvtest/nfvslaveservice/VmWorker.java
@@ -141,17 +141,17 @@ public class VmWorker extends MessageProcessor{
 				}
 				
 				//upload base vm image files
-				/*for(int i=0; i<chainConfig.stages.size(); i++){
+				for(int i=0; i<chainConfig.stages.size(); i++){
 					baseImgList.add(chainConfig.getImgNameForStage(i));
 					if(!agent.fileExistInDir(hostServer.hostServerConfig.imgDir, chainConfig.getImgNameForStage(i))){
 						String imgPath = hostServer.controllerConfig.imgDir+"/"+chainConfig.getImgNameForStage(i);
 						String remotePath = hostServer.hostServerConfig.imgDir+"/"+chainConfig.getImgNameForStage(i);
 						agent.uploadFile(imgPath, remotePath);
 					}
-				}*/
+				}
 				
 				//create management network
-				/*if(agent.networkExist(chainConfig.getManagementNetwork())){
+				if(agent.networkExist(chainConfig.getManagementNetwork())){
 					agent.deleteNetwork(chainConfig.getManagementNetwork());
 				}
 				String localMNetXMLFile = constructNetworkXmlFile(hostServer.controllerConfig,
@@ -160,10 +160,10 @@ public class VmWorker extends MessageProcessor{
 				String remoteMNetXMLFile = hostServer.hostServerConfig.xmlDir+"/"+
 										   chainConfig.getManagementNetwork();
 				agent.uploadFile(localMNetXMLFile, remoteMNetXMLFile);
-				agent.createNetworkFromXml(remoteMNetXMLFile);*/
+				agent.createNetworkFromXml(remoteMNetXMLFile);
 				
 				//create operational network if it's control plane
-				/*if(chainConfig.getOperationNetwork()!="nil"){
+				if(chainConfig.getOperationNetwork()!="nil"){
 					if(agent.networkExist(chainConfig.getOperationNetwork())){
 						agent.deleteNetwork(chainConfig.getOperationNetwork());
 					}
@@ -174,14 +174,14 @@ public class VmWorker extends MessageProcessor{
 											   chainConfig.getOperationNetwork();
 					agent.uploadFile(localMNetXMLFilz, remoteMNetXMLFilz);
 					agent.createNetworkFromXml(remoteMNetXMLFilz);
-				}*/
+				}
 			}
 			agent.createRouteToGateway(hostServer.entryIp, hostServer.gatewayIp, "gateway");
 			//remove previous generated vm images
-			/*String[] unusedImgArray = agent.createSelectedRemoveList(hostServer.hostServerConfig.imgDir, baseImgList);
+			String[] unusedImgArray = agent.createSelectedRemoveList(hostServer.hostServerConfig.imgDir, baseImgList);
 			for(int i=0; i<unusedImgArray.length; i++){
 				agent.removeFile(hostServer.hostServerConfig.imgDir+"/"+unusedImgArray[i]);
-			}*/
+			}
 			
 			agent.disconnect();
 		}
@@ -198,11 +198,11 @@ public class VmWorker extends MessageProcessor{
 		//set remove file names
 		logger.info("start creating new VM");
 		VmInstance vmInstance = request.getVmInstance();
-		/*String localXmlFile = constructLocalXmlFile(vmInstance);
+		String localXmlFile = constructLocalXmlFile(vmInstance);
 		String remoteXmlFile = vmInstance.hostServerConfig.xmlDir+"/"+vmInstance.vmName;
 		String remoteImgFile = vmInstance.hostServerConfig.imgDir+"/"+vmInstance.vmName;
 		String remoteBaseImgFile = vmInstance.hostServerConfig.imgDir+"/"+
-		                   vmInstance.serviceChainConfig.getImgNameForStage(vmInstance.stageIndex);*/
+		                   vmInstance.serviceChainConfig.getImgNameForStage(vmInstance.stageIndex);
 		logger.info("finish constructing xmls");
 		//this is the actual procedure to create vm on remote host server
 		HostAgent agent = new HostAgent(vmInstance.hostServerConfig);
@@ -215,7 +215,7 @@ public class VmWorker extends MessageProcessor{
 			//finally query the ovs ports for the vm and 
 			//set ovs ports in VmInstance
 			//reply to the actor that sends the CreateVmRequest
-			/*logger.info("start creating node "+vmInstance.managementIp);
+			logger.info("start creating node "+vmInstance.managementIp);
 			agent.connect();
 			agent.uploadFile(localXmlFile, remoteXmlFile);
 			for(int i=0; i<10; i++){
@@ -247,27 +247,10 @@ public class VmWorker extends MessageProcessor{
 			logger.info("finish creating node "+vmInstance.managementIp);
 			CreateVmReply reply = new CreateVmReply(this.getId(), request, true);
 			this.mh.sendTo(reply.getRequest().getSourceId(), reply);
-			Thread.sleep(5*1000);*/
+			Thread.sleep(5*1000);
 			
-			//agent connects to host server through ssh
-			//upload vm xml file to host server
-			//create vm image by copying base image
-			//create vm from the new xml file
-			//finally query the ovs ports for the vm and 
-			//set ovs ports in VmInstance
-			//reply to the actor that sends the CreateVmRequest
-			logger.info("start creating node "+vmInstance.managementIp);
+			/*logger.info("start creating node "+vmInstance.managementIp);
 			agent.connect();
-			//agent.uploadFile(localXmlFile, remoteXmlFile);
-			//agent.copyFile(remoteBaseImgFile, remoteImgFile);
-			//Thread.sleep(1000);
-			//for(int i=0; i<10; i++){
-			//	boolean successful = agent.createVMFromXml(remoteXmlFile);
-			//	if(successful == true){
-			//		logger.info("the node "+vmInstance.managementIp+" is successfully created");
-			//		break;
-			//	}
-			//}
 			int[] portList = new int[vmInstance.macList.size()];
 			if(vmInstance.serviceChainConfig.nVmInterface == 3){
 				int stageIndex = vmInstance.stageIndex;
@@ -292,20 +275,10 @@ public class VmWorker extends MessageProcessor{
 			else{
 				vmInstance.setPort(portList);
 			}
-			//int[] portList = new int[vmInstance.macList.size()];
-			//for(int i=0; i<vmInstance.macList.size(); i++){
-			//	String mac = vmInstance.macList.get(i);
-			//	String portMac = "fe:"+mac.substring(3);
-			//	int portNum = agent.getPort(vmInstance.macBridgeMap.get(mac), 
-			//			                    portMac);
-			//	portList[i] = portNum;
-			//}
-			//vmInstance.setPort(portList);
 			agent.disconnect();
 			logger.info("finish creating node "+vmInstance.managementIp);
 			CreateVmReply reply = new CreateVmReply(this.getId(), request, true);
-			this.mh.sendTo(reply.getRequest().getSourceId(), reply);
-			//Thread.sleep(5*1000);
+			this.mh.sendTo(reply.getRequest().getSourceId(), reply);*/
 		}
 		catch (Exception e){
 			e.printStackTrace();
